Remove libuv on Windows
Replace the libuv backend with an in-tree IOCP/AFD event-loop engine
implementing the us_loop_* contract, a native fd table (bit-tagged
indices, CRT lpReserved2 interop), NT-path file I/O with a worker-pool
async layer, named-pipe/ConPTY/tty/signal/fs-event handles, and
work-pool DNS. The uv symbol surface for N-API addons remains exported
(crash stubs + real polyfills); process.versions.uv reports the
emulated ABI level. POSIX behavior unchanged.

+++ src/runtime/cli/bunx_command.rs
@@ -1353,7 +1353,7 @@ impl BunxCommand {
                 }
 
                 if exited.code != 0 {
-                    Global::exit(exited.code as u32);
+                    Global::exit(exited.code);
                 }
             }
             SpawnStatus::Signaled(sig) => {
